# golang-backend/docs/swagger.yaml
basePath: /
definitions:
  controller.BaseSwaggerResp:
    properties:
      code:
        example: 0
        type: integer
      msg:
        example: 操作成功
        type: string
      ts:
        example: 1700000000000
        type: integer
    type: object
...
  controller.SwaggerConfigGetReq:
    properties:
      name:
        example: app_name
        type: string
    type: object
  controller.SwaggerConfigGetResp:
    properties:
      code:
        example: 0
        type: integer
      data:
        example: some-value
        type: string
      msg:
        example: 操作成功
        type: string
      ts:
        example: 1700000000000
        type: integer
    type: object
...
  controller.SwaggerConfigUpdateMap:
    additionalProperties:
      type: string
    type: object
  controller.SwaggerConfigUpdateSingle:
    properties:
      name:
        example: app_name
        type: string
      value:
        example: network-panel
        type: string
    type: object
...
  controller.SwaggerNodeExitReq:
    properties:
      ifaces:
        items:
          type: string
        type: array
      interface:
        type: string
      limiter:
        example: 5mbps
        type: string
      metadata:
        example: '{"k":"v"}'
        type: string
      method:
        example: AEAD_CHACHA20_POLY1305
        type: string
      nodeId:
        example: 1
        type: integer
      observer:
        example: console
        type: string
      password:
        example: pass
        type: string
      port:
        example: 10000
        type: integer
      rlimiter:
        example: ""
        type: string
    type: object
...
  controller.SwaggerNodeSimpleReq:
    properties:
      nodeId:
        example: 1
        type: integer
    type: object
...
  controller.SwaggerResp:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      ts:
        type: integer
    type: object
...
info:
  contact: {}
  description: 面板后端接口文档
  title: network-panel API
  version: "1.0"
paths:
  /api/v1/config/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 配置名
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerConfigGetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerConfigGetResp'
      summary: 根据 name 获取配置
      tags:
      - config
...
  /api/v1/config/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 键值对
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerConfigUpdateMap'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 批量更新配置
      tags:
      - config
  /api/v1/config/update-single:
    post:
      consumes:
      - application/json
      parameters:
      - description: name/value
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerConfigUpdateSingle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 单项配置更新
      tags:
      - config
...
  /api/v1/node/get-exit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 节点ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeSimpleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerResp'
      summary: 获取出口节点配置
      tags:
      - node
...
  /api/v1/node/set-exit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 出口服务配置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerNodeExitReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 配置出口节点服务
      tags:
      - node
...
swagger: "2.0"